requirements.txt
cat > requirements.txt << 'EOF'
flask==2.3.3
flask-cors==4.0.0
psycopg2-binary==2.9.7
redis==5.0.1
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
pandas==2.1.3
folium==0.15.0
polyline==2.0.2
openpyxl==3.1.2
xlrd==2.0.1
<<<<<<< HEAD
EOF
=======
>>>>>>> b4464ec (Auto-sync: 2026-03-20 16:59:19)
